Added support for barh. Cleaned up process for making bar collections.
This is the process for getting a bar chart to work:

1. Exporter calls draw_path(renderer).

2. PlotlyRenderer decides whether the path is a bar in a regular or
horizontal bar chart.

3. If the path is determined to be a bar, it is compared to all of the
other collections of bars that have been found.

4. If the path matches a collection, it's added. Else, a new collection
is created with the new path as the sole entry.

5. hbar charts are sorted vertically, bar charts are sorted
horizontally.

6. If a bar chart has more than one member, it is plotted. This is a
rudimentary method to prevent plotting nonsense bar charts. Else, the
bar chart is NOT plotted and a warning is raised explaining what
happened.

Currently, a single bar can be sorted into both hbar AND bar charts. The
hope is that one of the two collections will only include that single
bar and it will not be plotted.

In addition. There is no way for plotly to currently discern the
difference between one large bar chart and a chart that is meant to be
grouped bars. However, results don't end up looking all that bad and can
usually be changed at a high-level in GUIs on the plotly site.