Included makefile for build/install. Changed setup.py.
The setup.py module now imports the mplexporter subpackage within
matplotlylib. It also imports the mplexporter/renderers/base.py file.

The makefile is setup to allow users to sync changes with the
mplexporter submodule, sync the current version of the mplexporter
submodule with the one packaged within the matplotlylib package, and
install the matplotlylib package.