Added PATH_MAP to map Bezier curves to symbols. Also added size support.
Because of redundancies in the 'path codes', the map is many-to-one
instead of one-to-one. For instance the symbols, '^' '>' '<' 'v', all
map to the same Up Triangle symbol in plotly. This is because their
paths are identical: ['M', 'L', 'L', 'Z'].

With more work, the path vertices can be used to differentiate between
these.

Also, the marker size is crudely exported from the Affine2D mpl
transform object that mplexporter yields to PlotlyRenderer.