Updated renderer to deal with `draw_marked_line`. Updated tests.
This commit brings the rest of the matplotlylib project up to speed with
the mplexporter update that was just committed.

Now mpl Line2D objects that have markers and lines associated with them
can be treated as a single object by renderers, if desired.

Nosetests required updating to deal with the new syntax. On that note,
`label` and `size` were added to data dictionaries.