Added plotly_objs and plotly_words modules for better oop code.
Each portion of the plotly figure is now represented by an object which
inherit from either the standard `list` or `dict` classes.

The two major classes are the PlotlyDict and the PlotlyList.
Implementation of these objects allows for a more compartmentalized
approach that will be more robust to future changes. See the
documentation for more explanation.